Applies nominate test to L2
