+++ test/integration/l2/Owned.l2.integration.js
@@ -1,0 +1,9 @@
+const { bootstrapL2 } = require('../utils/bootstrap');
+const { itCanNominate } = require('../behaviors/nominate.behavior');
+
+describe('Owned integration tests (L2)', () => {
+	const ctx = this;
+	bootstrapL2({ ctx });
+
+	itCanNominate({ ctx });
+});
